refactor: split capability.rs into capability/ modules
2319 lines exceeded the codebase's file-size convention. Split into
capability/{state,tier,probe,diagnostics,guard}.rs with mod.rs as thin
re-export glue, following the config/ and storage/memory/ split pattern.
No behavior change; module boundaries follow the file's existing logical
sections (data types, the Tier enum, the probing engine, TLS diagnostics,
feature-gating guards).

Also: rename require_team_server to require_explicit_server_url (it works
identically against a self-hosted team server or Spelunk Cloud, so "team"
was misleading), and tighten its docstring to state plainly that every
current caller already established reachability via require_tier1 before
this guard runs, rather than emphasizing an "even offline" property that's
unexercised in practice.

Updates every comment across the crate that pointed at the old
crates/spelunk-cli/src/capability.rs path.
