--- docs/self-hosting.md
@@ -1,5 +1,0 @@
-# Self-hosting spelunk-server
-
-This page has moved. The team-server runbook (certificate, routable TLS bind,
-systemd unit, Docker, client trust configuration) now lives at
-[Server setup](server-setup.md).
