--- docs/server.md
@@ -1,9 +1,0 @@
-# spelunk-server
-
-This page has moved.
-
-- The local, zero-setup inference server is covered in
-  [Getting started](getting-started.md).
-- Deploying and configuring a shared **team server** (Docker, systemd, TLS,
-  client configuration, trust model) is now [Server setup](server-setup.md).
-- The HTTP API reference lives at [Server API](architecture/server-api.md).
